Fixed tests.

+++ tests/test_PSD.cpp
@@ -60,7 +60,7 @@ TEST_CASE("Calculate PSD @ 200 Hz sampling", "[PSD]") {
         CHECK(isnan(psd.accelerationPSDZ[i]));
     }
 
-    for (float i = 0; i < psd.numDataPoints * 2; i++) {
+    for (int i = 0; i < psd.numDataPoints * 2; i++) {
         double theta = 2 * M_PI * i / psd.numDataPoints;
         psd.append(1.0 * sin(5 * theta) + 0.5 * cos(13 * theta), 1.0 * cos(40 * theta) + 0.5 * sin(2 * theta),
                    1.0 * sin(13.6 * theta) + 0.5 * cos(10.2 * theta) + 0.4 * sin(22 * theta), false);
@@ -92,7 +92,7 @@ TEST_CASE("Calculate PSD @ 50 Hz sampling", "[PSD]") {
         CHECK(isnan(psd.accelerationPSDZ[i]));
     }
 
-    for (float i = 0; i < psd.numDataPoints * 2; i++) {
+    for (int i = 0; i < psd.numDataPoints * 2; i++) {
         double theta = M_PI * i * 0.02;
         psd.append(1.0 * sin(5 * theta) + 0.5 * cos(13 * theta), 1.0 * cos(40 * theta) + 0.5 * sin(2 * theta),
                    1.0 * sin(13.6 * theta) + 0.5 * cos(10.2 * theta) + 0.4 * sin(22 * theta), false);
